Added blank XML files for the series (no traces yet).
This would be the start of a new tracing project. These are the XML
files that reference images in another directory. In this case, the
images are expected to be "up" one directory and "down" into an "image"
directory and "down" again into the "even" directory. This looks like:

  project_directory
   |
   +-- series  (folder containing these XML files)
   |
   +-- images
        |
        +-- even  (folder containing the "even" images)

Inside each of the section files ("mystery_series.#") the images are
referenced with a relative path ("../images/even/mystery_#.jpeg"). This
was done to simulate a situation where the images are NOT part of the
Git repository that stores the XML files. Those images might be located
anywhere available to the system (possibly even on a shared resource).

What does matter, however, is that the relative path to the images is
the SAME for everyone who checks out the repository (even when the
images are not in the repository itself). This can be accomplished
fairly easily on UNIX-like systems with soft links. There may be similar
mechanisms on Windows machines. But either way, the images would not
have to be stored in the Git database.

Having said all of that, this repository does include both the images
and the XML files to make it easy for people to try it out. People
should be able to pull from this repository, open the series file in
Reconstruct, add or modify traces, and push the results back to the Git
repository. If there are conflicts, it means that someone else was
working on the same files at the same time. This will likely require a
manual merge. Part of this experiment is to find out how hard that will
be to do.

The typical workflow goes as follows:

  1. Pull changes before getting started
      (this will bring your local database up to date with others)

  2. Make modifications (add, delete, modify traces)

  3. Commit changes to your local Git repository (repeat if desired)

  4. Push all local commits to main Git repository to share

If people are working on separate sections, the only conflicts might be
in the series (.ser) files. In that case, the "merge strategy" might be
to just use the most recent version of the series file.

This will still require coordination between people working in the same
series, but hopefully the conflicts will be minimal. That's what this
test case is designed to find out.