Added traces to first 2 sections of the Mystery Series.
This is a test of using Reconstruct to add traces to 2 of the sections.

As expected, only 3 files were changed:

  series/mystery_series.1
  series/mystery_series.2
  series/mystery_series.ser

As mentioned in an earlier commit, the series file (.ser) will likely
change every time the series is opened. This is because it stores a
number of settings that can change due to viewing. It's not clear the
best way to handle this other than to just let it change in the Git
repository.