Removed Git change markers and other changes to sections 3,7,8
When a merge fails, Git will put edit "helpers" into the file itself to
mark the parts that Git didn't know what to do with.

These marks should be removed to fix the merge. If not, then the files
will contain both parts of the data (along with the git edit "helpers").