Removed lots of Git conflict markers.
When Git finds a conflict, it will mark the actual file with symbols:

  >>>>>>>>
  ========
  <<<<<<<<

These separate the differences between the two versions. These conflicts
should be resolved in favor of one version or the other (or a mixture)
before files are pushed to the repository. This may require some settings
on Git clients to not allow pushing when files are in that state.

It's interesting to note that the "C" version of Reconstruct doesn't
seem to mind certain kinds of errors in the XML files. Files with these
markers were able to be opened by that version. However, the Java version
is very strict and it will detect malformed XML and report an error.
This can be a helpful tool for finding these kinds of conflicts.

The good news is that Git keeps a history of everything that's been
committed and pushed. So these kinds of errors can generally be fixed
without losing any tracing information. In this case (a practice case),
it wasn't worth too much effort to save every trace, and some were just
deleted from the XML without considering their value. Also note that at
least one of the sections currently has a duplicate trace.

These are all the kinds of problems that are worth this exploration.