Removed Git's edit markers in Sections 16 through 29
There may be a misconfiguration of some of the Git tools because these
markers shouldn't be left in files that are committed.

These markers show where text has been inserted or remved from the
previous version, and they are placed in the text by Git to show the
decision that needs to be made by the person making the commit.